requirements.txt
anyio==4.12.1
boto3==1.42.63
botocore==1.42.63
certifi==2026.2.25
charset-normalizer==3.4.5
dnspython==2.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jmespath==1.1.0
pymongo==4.16.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
regex==2026.2.28
requests==2.32.5
s3transfer==0.16.0
six==1.17.0
tavily-python==0.7.22
tiktoken==0.12.0
urllib3==2.6.3
google-genai==1.68.0
boto3==1.34.0
